Handle pasting without codemirror
Stops codemirror from handling the event in order to prevent extra text from showing up when pasting an image or other file. Then extracts the items with type text/plain to avoid pasting html and checks that the next item in the clipboard is not a file. If the next item is a file then we do not want to paste the text as it will be a path to the file.